+++ Libraries/01-Shared/Elysium.Core.Security/StoreName.hpp
@@ -19,7 +19,7 @@ Copyright (C) 2017 waYne (CAM)
 namespace Elysium::Core::Security::Cryptography::X509Certificates
 {
 #if defined(_WIN32) || defined(_WIN64) || defined(_WINDOWS)
-	enum class StoreName : uint32_t
+	enum class StoreName : uint8_t
 #elif defined(__ANDROID__)
 	enum class StoreName
 #else
